socialbloggingcrew/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.1.4
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
backoff==2.2.1
bcrypt==4.3.0
blinker==1.7.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2023.11.17
chardet==5.2.0
chromadb==1.0.15
click==8.1.6
cloud-init==25.1.4
colorama==0.4.6
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
dbus-python==1.3.2
distro==1.9.0
distro-info==1.7+build1
durationpy==0.10
fastapi==0.116.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.7.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.20.4
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==22.10.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
netifaces==0.11.0
numpy==2.3.2
oauthlib==3.2.2
onnxruntime==1.22.1
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.1
overrides==7.7.0
packaging==24.0
pipx==1.4.3
platformdirs==4.2.0
posthog==5.4.0
protobuf==6.31.1
psutil==5.9.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybase64==1.4.2
pycurl==7.45.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.36.2
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.26.0
rsa==4.9.1
service-identity==24.1.0
setuptools==68.1.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
systemd-python==235
tenacity==9.1.2
tokenizers==0.21.4
tqdm==4.67.1
Twisted==24.3.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==2.0.7
userpath==1.9.1
uv==0.8.4
uvicorn==0.35.0
uvloop==0.21.0
wadllib==1.3.6
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wheel==0.42.0
zipp==3.23.0
zope.interface==6.1
